feat: added technologies section
