+++ src/components/main-layout.tsx
@@ -2,14 +2,14 @@ import React, { ReactNode } from 'react';
 import { ThemeProvider, createTheme } from '@mui/material/styles';
 
 const theme = createTheme({
-    palette: {
-        primary: {
-        main: '#4F37EA'
-        },
-        secondary: {
-        main: '#D714E8'
-        },
+  palette: {
+    primary: {
+      main: '#4F37EA'
     },
+    secondary: {
+      main: '#D714E8'
+    },
+  },
 });
 
 const MainLayout: React.FC<{ children: ReactNode }> = ({ children }) => {
