+++ src/images/partners/Levvy.tsx
@@ -1,0 +1,83 @@
+import React from 'react'
+
+const Levvy = () => {
+    return (<div className="mr-auto">
+        <svg width="200" height="90" viewBox="0 0 128 32" fill="none" xmlns="http://www.w3.org/2000/svg">
+            <g clipPath="url(#clip0_62_778)">
+                <g clipPath="url(#clip1_62_778)">
+                    <g clipPath="url(#clip2_62_778)">
+                        <path
+                            d="M82.0014 8.1511C82.0014 7.19719 82.434 6.35032 83.1146 5.81189C83.2469 5.70802 83.2146 5.50028 83.0604 5.43457C82.7122 5.28724 82.3256 5.20669 81.9097 5.20669C80.7673 5.20669 79.6625 5.9041 79.2435 7.06576L75.7122 17.4348C75.6434 17.6266 75.3766 17.6277 75.3067 17.4348L71.7389 7.0647C71.3199 5.90304 70.2151 5.20563 69.0728 5.20563C67.4728 5.20563 66.2158 6.52309 66.2158 8.15004C66.2158 8.49875 66.3305 9.47386 66.4441 9.82257L71.4346 22.3347C72.0819 23.9617 73.6443 25.1626 75.511 25.1626C77.3778 25.1626 78.9391 23.9617 79.5864 22.3347L83.3564 12.8104C83.3772 12.7574 83.3772 12.6981 83.3564 12.6462L82.2307 9.27354C82.1161 8.92483 82.0024 8.49875 82.0024 8.15004L82.0014 8.1511Z"
+                            fill="url(#paint0_linear_62_778)" />
+                        <path
+                            d="M102 8.15108C102 6.52413 100.743 5.20667 99.1044 5.20667C97.9235 5.20667 96.8187 5.90408 96.4383 7.06574L92.9069 17.7294C92.8392 17.9234 92.5692 17.9245 92.5005 17.7305L88.8962 7.06574C88.5157 5.86487 87.4109 5.20667 86.23 5.20667C84.6301 5.20667 83.3731 6.52413 83.3731 8.15108C83.3731 8.49979 83.4877 8.92587 83.6399 9.27458C83.717 9.35937 87.6517 18.8434 89.0088 23.1477C89.0234 23.1932 89.0369 23.2367 89.0505 23.2802C89.1245 23.4921 89.1859 23.6935 89.2287 23.879C89.2766 24.0878 89.3037 24.2786 89.3037 24.445C89.3037 25.5685 88.8368 26.6983 87.4745 26.6983C86.1122 26.6983 85.6067 25.8462 84.5915 25.8462C83.0468 25.8462 82.2734 26.6146 82.2734 28.4737C82.2734 31.0302 84.9553 31.9989 87.191 31.9989C88.128 31.9989 88.9827 31.824 89.7634 31.5135C93.4562 30.0455 95.5064 25.543 96.7812 22.1217C97.0344 21.4211 97.2981 20.7671 97.566 20.128C97.6244 19.9574 101.734 9.27458 101.734 9.27458C101.849 8.92587 102.001 8.49979 102.001 8.15108H102Z"
+                            fill="url(#paint1_linear_62_778)" />
+                        <path
+                            d="M61.7112 21.9871C61.6559 22.0496 61.5996 22.1121 61.5434 22.1736C61.4308 22.2966 61.3151 22.4153 61.1984 22.5319C60.7439 22.9855 60.2593 23.3957 59.7454 23.7645C59.4056 24.0083 59.0523 24.2351 58.6875 24.4429C58.455 24.5754 58.2174 24.7015 57.9745 24.8202C57.8057 24.9029 57.8609 25.1636 58.0485 25.1636C58.0506 25.1636 58.0537 25.1636 58.0558 25.1636C59.884 25.1636 61.4839 23.9627 62.1312 22.3358L62.7952 20.658C62.8097 20.6219 62.7608 20.5944 62.7389 20.6262C62.5002 20.9696 62.0447 21.6087 61.7112 21.986V21.9871Z"
+                            fill="url(#paint2_linear_62_778)" />
+                        <path
+                            d="M64.7745 9.82365C64.6598 9.47494 64.5462 8.49983 64.5462 8.15112C64.5462 7.1972 64.9788 6.35034 65.6594 5.81191C65.7918 5.70804 65.7595 5.5003 65.6052 5.43564C65.2571 5.28832 64.8704 5.20776 64.4545 5.20776C63.3122 5.20776 62.2073 5.90518 61.7883 7.06684L59.9101 12.8401C59.8392 13.0373 60.055 13.2143 60.2301 13.1019C60.4271 12.9758 60.6209 12.8412 60.8096 12.6992C61.1817 12.4194 61.5361 12.1088 61.8717 11.7633C61.8988 11.7294 61.9249 11.7029 61.9509 11.6827C61.977 11.6626 62.0009 11.6488 62.0239 11.6393C62.0395 11.6329 62.0541 11.6297 62.0677 11.6265C62.0885 11.6234 62.1083 11.6234 62.125 11.6265C62.1531 11.6308 62.1761 11.6414 62.1906 11.6509C62.199 11.6562 62.2063 11.6647 62.2136 11.6742C62.2157 11.6764 62.2167 11.6785 62.2188 11.6817C62.2334 11.7029 62.2438 11.7304 62.2501 11.7654C62.2501 11.7696 62.2511 11.7728 62.2521 11.777C62.2542 11.7951 62.2553 11.8152 62.2542 11.8364C62.2521 11.8671 62.2459 11.9011 62.2344 11.9371C62.1844 12.0918 62.1625 12.1724 62.0781 12.3091C61.4037 13.4093 60.4771 14.2593 59.4244 14.9599C58.4029 15.6404 56.2454 16.8868 52.3597 17.527C52.2169 17.5503 52.1262 17.6913 52.1804 17.828L52.9017 19.6373C52.9361 19.7242 53.0205 19.7783 53.1122 19.774C53.1404 19.773 53.1685 19.7708 53.1977 19.7698C56.9531 19.5578 60.9357 18.0315 63.1527 14.7395C63.1621 14.7268 63.1725 14.7151 63.186 14.7045C63.209 14.6865 63.235 14.6769 63.2611 14.6695C63.2892 14.6621 63.3174 14.66 63.3445 14.6642C63.3476 14.6642 63.3518 14.6642 63.3559 14.6642C63.3935 14.6716 63.4299 14.6886 63.4612 14.713C63.4925 14.7374 63.5175 14.7702 63.5331 14.8084C63.5415 14.8275 63.5467 14.8476 63.5498 14.8699C63.5498 14.8762 63.5498 14.8815 63.5498 14.8879C63.5498 14.9027 63.5498 14.9165 63.5467 14.9313C63.5352 14.9833 63.5206 15.0352 63.506 15.0861C63.5008 15.103 63.4966 15.121 63.4914 15.138C63.48 15.1772 63.4664 15.2164 63.4539 15.2557C63.4424 15.2906 63.4299 15.3245 63.4174 15.3585C63.3809 15.456 63.3424 15.5535 63.3007 15.6489C62.8608 16.6526 61.5496 17.8959 60.7658 18.6368C59.1982 20.1185 57.3513 21.0957 55.299 21.6797C54.905 21.7921 54.5089 21.8854 54.1118 21.9564C53.9753 21.9807 53.894 22.1249 53.9461 22.2553L53.9784 22.3358C54.1879 22.8615 54.4922 23.3427 54.8716 23.7529C54.9206 23.8059 54.9936 23.8303 55.0645 23.8186C55.6273 23.7264 56.2037 23.6236 56.7707 23.4159C63.505 20.941 65.9002 12.8115 65.9002 12.8115C65.921 12.7585 65.921 12.6992 65.9002 12.6472L64.7745 9.82471V9.82365Z"
+                            fill="url(#paint3_linear_62_778)" />
+                        <path
+                            d="M56.7404 14.377L54.2848 7.06684C53.8658 5.90518 52.7609 5.20776 51.6186 5.20776C50.2938 5.20776 49.2067 5.76845 48.8711 6.92375C50.6378 8.5719 51.6905 10.8539 51.6905 13.3118C51.6905 14.023 51.5519 14.5858 51.2871 15.0214C51.2986 15.1348 51.3914 15.226 51.5081 15.2345C53.4186 15.3659 54.4974 15.2313 56.5924 14.6674C56.7154 14.6346 56.7853 14.4999 56.7404 14.3781V14.377Z"
+                            fill="url(#paint4_linear_62_778)" />
+                        <path
+                            d="M30.5193 19.5737H27.7676C27.6478 19.5737 27.5519 19.4751 27.5519 19.3543V8.3705C27.5519 6.54958 26.1041 5 24.2759 5C22.4477 5 21 6.54958 21 8.3705V21.6394C21 23.5865 22.552 25.1647 24.4667 25.1647H35.9914C36.3343 25.1647 36.3541 24.7821 36.0289 24.6718C35.0001 24.321 32.6206 23.3045 31.2166 20.022C31.0978 19.7443 30.8174 19.5747 30.5193 19.5747V19.5737Z"
+                            fill="url(#paint5_linear_62_778)" />
+                        <path
+                            d="M47.5506 7.54589C45.9611 5.98465 43.7535 5 41.203 5C35.709 5 31.8525 9.40284 31.8525 14.856C31.8525 20.9728 36.0363 24.7513 41.4573 24.7513C43.3126 24.7513 45.2408 24.3613 46.9512 23.4657C47.6788 23.1148 48.2249 22.3358 48.2249 21.401C48.2249 20.1153 47.2785 19.102 46.1143 19.102C45.787 19.102 45.4597 19.1805 45.2044 19.3363C44.2225 19.8429 43.1302 20.1153 42.1483 20.1153C39.5291 20.1153 38.001 18.518 37.6738 16.7649H46.8418C48.2854 16.7649 49.3006 16.412 49.8113 15.528C50.0594 15.0988 50.1886 14.5444 50.1886 13.8428C50.1886 11.4198 49.2037 9.16966 47.5495 7.54483L47.5506 7.54589ZM37.6018 12.8698C37.8197 11.5057 38.7661 9.63602 41.204 9.63602C43.7868 9.63602 44.6968 11.5449 44.8427 12.8698H37.6029H37.6018Z"
+                            fill="url(#paint6_linear_62_778)" />
+                    </g>
+                </g>
+            </g>
+            <defs>
+                <linearGradient id="paint0_linear_62_778" x1="22.3184" y1="25.5134" x2="100.149"
+                    y2="9.97902" gradientUnits="userSpaceOnUse">
+                    <stop stopColor="#AD622D" />
+                    <stop offset="1" stopColor="#38987F" />
+                </linearGradient>
+                <linearGradient id="paint1_linear_62_778" x1="23.5504" y1="31.6841" x2="101.381"
+                    y2="16.1498" gradientUnits="userSpaceOnUse">
+                    <stop stopColor="#AD622D" />
+                    <stop offset="1" stopColor="#38987F" />
+                </linearGradient>
+                <linearGradient id="paint2_linear_62_778" x1="23.3307" y1="30.585" x2="101.161" y2="15.0507"
+                    gradientUnits="userSpaceOnUse">
+                    <stop stopColor="#AD622D" />
+                    <stop offset="1" stopColor="#38987F" />
+                </linearGradient>
+                <linearGradient id="paint3_linear_62_778" x1="21.687" y1="22.3507" x2="99.5177" y2="6.81628"
+                    gradientUnits="userSpaceOnUse">
+                    <stop stopColor="#AD622D" />
+                    <stop offset="1" stopColor="#38987F" />
+                </linearGradient>
+                <linearGradient id="paint4_linear_62_778" x1="20.5863" y1="16.836" x2="98.4181" y2="1.30263"
+                    gradientUnits="userSpaceOnUse">
+                    <stop stopColor="#AD622D" />
+                    <stop offset="1" stopColor="#38987F" />
+                </linearGradient>
+                <linearGradient id="paint5_linear_62_778" x1="20.6227" y1="17.0161" x2="98.4534"
+                    y2="1.48176" gradientUnits="userSpaceOnUse">
+                    <stop stopColor="#AD622D" />
+                    <stop offset="1" stopColor="#38987F" />
+                </linearGradient>
+                <linearGradient id="paint6_linear_62_778" x1="21.0377" y1="19.0967" x2="98.8684"
+                    y2="3.56235" gradientUnits="userSpaceOnUse">
+                    <stop stopColor="#AD622D" />
+                    <stop offset="1" stopColor="#38987F" />
+                </linearGradient>
+                <clipPath id="clip0_62_778">
+                    <rect width="128" height="32" fill="white" />
+                </clipPath>
+                <clipPath id="clip1_62_778">
+                    <rect width="81" height="27" fill="white" transform="translate(21 5)" />
+                </clipPath>
+                <clipPath id="clip2_62_778">
+                    <rect width="81" height="27" fill="white" transform="translate(21 5)" />
+                </clipPath>
+            </defs>
+        </svg>
+    </div>);
+}
+
+export default Levvy;
